requirements.txt
Django==1.6.2
MySQL-python==1.2.5
South==0.8.4
beautifulsoup4==4.3.2
dj-database-url==0.3.0
dj-static==0.0.5
django-autocomplete-light==2.0.0a15
django-grappelli==2.5.2
django-toolbelt==0.0.1
django-zurb-foundation==5.1.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
six==1.6.1
static==1.0.2
wsgiref==0.1.2
